src/requirements.txt
discord.py==1.7.3
discord_py_slash_command==3.0.3
discord_py_interactions==4.1.0
Flask==2.2.2
pandas==1.5.1
promise==2.3
requests==2.28.1
